Add compatibility to aiohttp to 2.2.5 and bump version
aiohttp 2 added a maximum size to requests which we needed to increase for
the fake API server to upload large blobs. I've set the request size on the
callback server to 8192 bytes which should be sufficient for any incoming
data from the Threema gateway server. (Note that downloading large blobs is
unaffected by this size limitation.)